inventoryManagement/requirements.txt
djangorestframework 
mysqlclient 
python-decouple
djangorestframework-simplejwt
django >=4.0
